refactor: P2 server split 3/N — extract /api/settings + /api/mcp route domain
Pure cut-paste, zero behavior change:

- server/routes/settings.ts: all 19 blocks covering /api/settings* (model
  switch, provider CRUD + probe-models, channels, memory, simple-mode, mcp
  switch, github/ocr/tavily/content-hub tokens, theme, close-behavior) and
  /api/mcp* server management become handleSettingsRoutes(...):
  Promise<boolean>
- The domain-private helpers move with it: buildSafeSettings (secret
  redaction), maskSecret, providerModelToRuntimeModel, parseProviderPayload,
  parseModelConfig, parseStringHeaders
- Mutable config flows through getConfig/setConfig; a per-request local
  view keeps multi-save blocks (providers PUT) reading their own writes,
  mirrored to server.ts via setConfig. reloadFeishuChannel,
  scheduleSkillsReload and invalidateContentSource stay owned by server.ts
  (shared with other domains) and are injected as callbacks
- server.ts keeps one delegation call; drops imports that only the
  settings domain used (provider CRUD helpers, probeProviderModels, mcp
  store functions, getAvailableModels/switchModel/
  refreshConfiguredProviders)
- Smoke test gains three settings assertions (theme accept/reject, memory
  boolean validation, /api/mcp overview)

server.ts 5155 -> 4607 lines.
